+++ CHANGELOG.md
@@ -4,6 +4,10 @@ IMPROVEMENTS:
 * provider: Add `cdp_config_file`, `cdp_shared_credentials_file`, `cdp_endpoint_url` and `endpoint_url` arguments
 * resource/cdp_environments_aws_environment: Add retries with exponential backoff to ride over consistency issues.
 
+FEATURES:
+
+* **New Resource:** `cdp_environments_azure_credential`
+
 ## 0.0.2 (June 12, 2020)
 
 NOTES: